Optimize for consumption by npm-packlist
Cache the contents of node_modules folders and package.json files
so that packlist doesn't have to re-read.

Compile the list of dep names at the top level only.  There's no need
to note that nested deps are bundled, because they'll be included by
default when their bundled parent is included.